chore(inspect_ai_evals): update tasks list
