+++ jobs/inspect_ai_evals/api_model/sample-schema.json
@@ -23,18 +23,27 @@
           "agie_sat_en",
           "agie_sat_en_without_passage",
           "agie_sat_math",
+          "ahb",
           "aime2024",
+          "aime2025",
           "air_bench",
           "arc_challenge",
           "arc_easy",
+          "bbh",
           "bbeh",
           "bbeh_mini",
+          "bbq",
+          "bfcl",
           "bbh",
           "bbq",
+          "bfcl",
+          "bold",
           "boolq",
           "chembench",
+          "coconot",
           "commonsense_qa",
           "cyse3_visual_prompt_injection",
+          "docvqa",
           "drop",
           "gpqa_diamond",
           "gsm8k",
@@ -63,8 +72,10 @@
           "lab_bench_tableqa",
           "lingoly",
           "lingoly_too",
+          "make_me_pay",
           "mask",
           "math",
+          "mathvista",
           "medqa",
           "mgsm",
           "mmiu",
@@ -92,14 +103,18 @@
           "sevenllm_qa_en",
           "sevenllm_qa_zh",
           "simpleqa",
+          "simpleqa_verified",
           "sosbench",
+          "squad",
           "stereoset",
           "strong_reject",
           "sycophancy",
           "truthfulqa",
+          "uccb",
           "vstar_bench_attribute_recognition",
           "vstar_bench_spatial_relationship_reasoning",
           "winogrande",
+          "worldsense",
           "wmdp_bio",
           "wmdp_chem",
           "wmdp_cyber",
